MIC-4650: Refactor PaymentController
