+++ TinkoffASDKUI/TinkoffASDKUI/BusinessLogic/Payment/Controller/PaymentControllerDataSource.swift
@@ -1,0 +1,11 @@
+//
+//  PaymentControllerDataSource.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 19.02.2023.
+//
+
+import Foundation
+
+/// Объект, предоставляющий дополнительный данные для `PaymentController`
+public protocol PaymentControllerDataSource: AnyObject {}
