+++ TinkoffASDKUI/TinkoffASDKUI/BusinessLogic/Payment/Process/PaymentProcessDelegate.swift
@@ -1,22 +1,11 @@
 //
+//  PaymentProcessDelegate.swift
+//  TinkoffASDKUI
 //
-//  PaymentProcess.swift
-//
-//  Copyright (c) 2021 Tinkoff Bank
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
+//  Created by r.akhmadeev on 19.02.2023.
 //
 
+import Foundation
 import TinkoffASDKCore
 
 protocol PaymentProcessDelegate: AnyObject {
@@ -58,11 +47,4 @@ protocol PaymentProcessDelegate: AnyObject {
         confirmationCancelled: @escaping () -> Void,
         completion: @escaping (Result<GetPaymentStatePayload, Error>) -> Void
     )
-}
-
-public protocol PaymentProcess: Cancellable {
-    var paymentId: String? { get }
-    var paymentFlow: PaymentFlow { get }
-    var paymentSource: PaymentSourceData { get }
-    func start()
 }
