+++ TinkoffASDKUI/TinkoffASDKUI/BusinessLogic/WebBasedThreeDS/ThreeDSWebFlow/Controller/ThreeDSWebFlowDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  ThreeDSWebFlowDelegate.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 19.02.2023.
+//
+
+import Foundation
+import UIKit
+import WebKit
+
+/// Объект, предоставляющий UI-компоненты для прохождения `3DS Web Based Flow`
+public protocol ThreeDSWebFlowDelegate: AnyObject {
+    /// webView, в котором выполнится запрос для прохождения 3DSChecking
+    func hiddenWebViewToCollect3DSData() -> WKWebView
+    /// viewController для модального показа экранов, необходимость в которых может возникнуть в процессе оплаты
+    func sourceViewControllerToPresent() -> UIViewController?
+}
