+++ TinkoffASDKUI/TinkoffASDKUI/YandexPay/Implementations/YandexPayPaymentSheet/YandexPayPaymentSheetAssembly.swift
@@ -37,7 +37,7 @@ final class YandexPayPaymentSheetAssembly: IYandexPayPaymentSheetAssembly {
         presenter.view = view
         paymentControllerUIProvider.view = view
         paymentController.delegate = presenter
-        paymentController.uiProvider = paymentControllerUIProvider
+        paymentController.webFlowDelegate = paymentControllerUIProvider
 
         return PullableContainerViewController(content: view)
     }
